feat: :sparkles: creaing ray class to represent the camera ray
This commit introduces the Ray class, which is designed to represent a ray in 3D space. The Ray class is initialized with an origin and a direction, both of which are expected to be instances of the Ponto and Vetor classes respectively.  The Ray class includes the following methods: - `__init__`: Initializes the Ray instance with an origin and a direction. - `__str__` and `__repr__`: Return a string representation of the Ray instance. - `get_point`: Returns a point at a given distance from the origin along the direction of the ray. - `__add__`: Returns a new Ray instance that is the result of the component-wise addition of two Ray instances. - `__sub__`: Returns a new Ray instance that is the result of the component-wise subtraction of two Ray instances. - `__mul__`: Returns a new Ray instance that is the result of multiplying the components of a Ray instance by a scalar. - `__truediv__`: Returns a new Ray instance that is the result of dividing the components of a Ray instance by a scalar.  These methods allow for the manipulation and interaction of Ray instances in a manner consistent with their geometric interpretation.